+++ Headers/GraphicsResources/CircleOutlineGraphicsResource.h
@@ -13,7 +13,7 @@ class CircleOutlineGraphicsResource : public IGraphicsResource
 {
 private:
 
-    uint8_t radius;
+    uint16_t radius;
     uint8_t thickness;
 
     SDL_Surface* circleSurface;
@@ -22,7 +22,7 @@ public:
 
     CircleOutlineGraphicsResource(int id, std::string resourceName, float radius, float thickness, uint8_t a, uint8_t r, uint8_t g, uint8_t b) : IGraphicsResource(id, resourceName)
     {
-        this->radius = static_cast<uint8_t>(floor(radius));
+        this->radius = static_cast<uint16_t>(floor(radius));
 
         if ((thickness < radius) && (thickness > 0.0f)) {
             this->thickness = static_cast<uint8_t>(floor(thickness));
@@ -61,8 +61,8 @@ public:
 
         SDL_Texture* circleTexture = SDL_CreateTextureFromSurface(sdlRenderer, this->circleSurface);
 
-        uint8_t x = static_cast<int>(transformComponent->position.y) - this->radius;
-        uint8_t y = static_cast<int>(transformComponent->position.x) - this->radius;
+        uint16_t x = static_cast<uint16_t>(transformComponent->position.y) - this->radius;
+        uint16_t y = static_cast<uint16_t>(transformComponent->position.x) - this->radius;
 
         SDL_Rect dstRect = {x, y, 2*this->radius, 2*this->radius};
